Simplifies implementation of method of double class.
