--- src/test/java/client/FakeClientSocket.java
@@ -1,25 +1,0 @@
-package client;
-
-import java.io.ByteArrayOutputStream;
-import java.io.PrintWriter;
-import java.util.ArrayList;
-import java.util.List;
-
-public class FakeClientSocket implements WritingSocket {
-
-    public boolean isClosed;
-    public List<String> allMessagesWritten = new ArrayList<>();
-
-    @Override
-    public void writeToStream(String userInput) {
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        PrintWriter printWriter = new PrintWriter(outputStream, true);
-        printWriter.println(userInput);
-        allMessagesWritten.add(outputStream.toString());
-    }
-
-    @Override
-    public void close() {
-        isClosed = true;
-    }
-}
